move send_heartbeat_signal to webserver scheduler
