+++ services/orchest-webserver/app/app/core/scheduler.py
@@ -22,6 +22,7 @@ import enum
 import json
 import logging
 import os
+import time
 from typing import Callable
 
 import requests
@@ -115,10 +116,45 @@ class Jobs:
                 `interval=0` will execute the job right away.
 
         """
+        # Nested function because it won't be called from any other
+        # place.
+        def send_heartbeat_signal(app: Flask) -> None:
+            """Sends a heartbeat signal to the telemetry service.
+
+            A user is considered to be active if the user has triggered
+            any webserver logs in the last half of the
+            `TELEMETRY_INTERVAL`.
+
+            """
+            # Value of None indicates that the user's activity could not
+            # be determined.
+            active = None
+            try:
+                t = os.path.getmtime(app.config["WEBSERVER_LOGS"])
+            except OSError:
+                app.logger.error(
+                    (
+                        "Analytics heartbeat failed to identify "
+                        "whether the user is active."
+                    ),
+                    exc_info=True,
+                )
+            else:
+                diff_minutes = (time.time() - t) / 60
+                active = diff_minutes < (app.config["TELEMETRY_INTERVAL"] * 0.5)
+
+            analytics.send_event(
+                app, analytics.Event.HEARTBEAT_TRIGGER, {"active": active}
+            )
+            app.logger.debug(
+                "Successfully sent analytics event "
+                f"'{analytics.Event.HEARTBEAT_TRIGGER}'."
+            )
+
         return self._handle_recurring_scheduler_job(
             SchedulerJobType.TELEMETRY_HEARTBEAT.value,
             interval,
-            analytics.send_heartbeat_signal,
+            send_heartbeat_signal,
             app,
         )
